v.1.0.0

+++ Lab_5/PROGRAMM/include/N-body-problem-cuda.cuh
@@ -18,7 +18,8 @@
 
 
 // Используемый тип данных для вычислений
-typedef double mytype;
+//typedef double mytype;
+typedef float mytype;
